Reformatted comment lines to fit in 80 character width
I wished to have the text fit in a terminal window to allow comfortable
reading without having to set line wrapping.

I noticed several typos, duplicate/missing words and wording that could
be clarified. This is a preliminary step before applying each of these
changes, to ensure that each set of fixes can be reviewed more easily
(a change in line breaks results in many changes in diff viewers).