Clarification of the default culture
The plugin description uses "default culture" to refer both to
jQuery.culture, as in "the default culture if no culture is specified on
the various parsing and formatting functions", and to refer to
jQuery.cultures["default"] which is set to the neutral English culture
"en" at the start of jquery.glob.js, as in "if no match is found, the
'default' culture is selected.".

It became very confusing when I realized that calling
jQuery.preferCulture() would select the 'default' culture as default
culture by default when no match is found...

I made a couple of rewordings to reduce the ambiguity, trying to clarify
in which context each of jQuery.culture and jQuery.cultures.en is used
by default.

I have not updated all the mentions to the 'default' culture in
"Defining Culture Information", yet. Mostly because the code sample
explicitly assigns the English culture object to
jQuery.cultures['default']. But the actual code assigns to both
cultures["default"] and cultures.en at the same time.