Completed clarification of default culture in "Defining Culture..."
I rewrote large parts of the description in the section "Defining
Culture Information", so as to clarify how each culture is actually
defined as an extension of English.

This rather came as a shock to me, as I had safely assumed that each
culture (language+country) would extend the corresponding neutral
culture(language). There is lots of lots of duplicate information in
current cultures as defined.

I also updated the sample code to initialize both
jQuery.cultures['default'] and jQuery.cultures.en.