Simplified sample code in "jQuery.preferCulture"
After having read this a couple of times now, having the HTML markup at
the start of the code sample does not help at all. It takes some time to
figure out that the first line duplicates the description of the
sentence before "if the application has included only the neutral "fr"
culture".