Clarification in "jQuery.preferCulture" (3rd paragraph)
I added details about the expected array argument. After checking in the
code, the processing on each item of the array is the same as the
processing described for a single name string. I modified the phrase to:
"If you pass an array of names instead of a single name string".