Fixed and clarified "jQuery.findClosestCulture"
I clarified the description of jQuery.findClosestCulture, and in so doing,
found out that the assignment was described in the wrong way "setting jQuery
culture to it" rather than the actual behavior: "setting it to jQuery.culture".

For consistency and added precision, I also replaced "jQuery.culture" with
"the jQuery.culture property".