Added (fr) in French values in 3rd code sample of "jQuery.localize"
(en) was present in English values. Adding (fr) in French values made the code
sample much clearer, which allowed me to notice a cut and paste at the end of
the sample: the jQuery.culture property is reset to the French culture after
setting it to Spanish, in which case the result of the following line does not
make sense. It will fix that in the following commit.