Don't write to untranslated_attributes when changing a translated one
Writing to a translated attribute causes the untranslated attribute
with the same name to be modified, due to a rogue call to super in
write_attribute.

The parent method should only be called if the attribute is not
translated. If it is translated, to keep dirty objects tracking
working we just call attribute_will_change! before writing the
new value.

Many thanks to @pilif for the help in tracking this one down.