Fixes #44
The parseInt and parseFloat functions failed on negative currency
patterns with parentheses for cultures whose number formatting rules did
not contain parentheses (for example, "en"). The fix was to simply
attempt to parse the numbers using the pattern "(n)" if a sign was not
found using their default number pattern.