Add lifecycle setup/teardown methods to test suite
This ensures that every test beings with the default culture selected
and the original culture definitions intact. Introduces two global
variables: (1) lifecycle, which is an object containing the setup and
teardown functions for use in Qunit.module and (2) originalCultures,
which is a method that will return an unmodified version of a specific
culture definition (if an string argument is given) or all culture
definitions if called with no argument.