extend c format to allow specification of target currency
it is concievable that you need to emit a value in a currency foreign to
the current locale - let's say you have an online shop and need to show
the user (for example in en_US) a price in a different country (say,
EUR).

To use this, pass the target currency symbol in () after the c
specifier. Like this:

Globalize.format(1234.12, 'c(EUR)');