Test that cache_key changes when translation modified
As reported in https://github.com/globalize/globalize/pull/330 if a
translated model object is used as a cache key for a cached block in a
template, then the cache key should change if the translation of the
item changes.

This adds a unit test to verify this scenario. Note that I may have
misunderstood the original scenario, but I have verified that if the
`:touch => true` is removed from `#translation_class` then this test
fails.

I added a new `Product` model to support this test, because none of the
other models had timestamps on. The timestamps are needed to exhibit the
fact that the cache_key changes on modification of the
translation.

I considered adding timestamps to an existing model but I was concerned
about unintended side-effects, so I opted for adding a new model
instead.