Cache key incorporates translation cache key
This ensures that when the translation is modified that modification is
reflected in the cache key.

It means that if either the parent model or the translation is modified
(causing the `updated_at` timestamp to change) then the `cache_key` will
also change.

With this in place, we can remove `:touch => true` for the association
from the translation to the parent model, which was causing problems
with optimistic locking[1].

[1]
https://github.com/globalize/globalize/pull/330#issuecomment-50129369