Docs: Dedupe documentation for alias methods
Several methods have an alias. For example, formatNumber (a) and
numberFormatter (b).

a) `.formatNumber( value [, options] )`.
b) `.numberFormatter( [options] )( value )`.

Keeping duplicate documentation is error prone and confusing. In this
commit, the documentation is kept for one of the methods only
(formatters and parsers) and the documentation for the alias duplicate
is removed or merged.