Date: Parse February correctly in leap years.
When setting the month, the date parser may set the day to the last day
of the month.  If the year has not yet been set, this will be the last
day of the month in the current year.  So for example, in 2016 "2/1/15"
would be month-parsed to "Feb 29, 2016", then year-parsed to
"Mar 1, 2015".  The fix is set the month after the year has been set.

Also added a test for this case.

Fixes: 585.