Date: Fix parser lookup (use descending order)
parseDate fails on Turkish full datetime with Monday or Saturday.

The root cause is that the parser lookup finds Cuma (Fri) before it
finds Cumartesi (Sat), similarly it finds Pazar (Sun) before it finds
Pazartesi (Mon), and therefore the token doesn't entirely consume the
input.

The fix approach is to start the lookup with longer names, then the
shorter ones.

Fixes #690