Number: Remove unused code, fix strict mode
Comments from rxaviers:

The variable name `negativeSuffix` was a typo. It should have been 
`negativeProperties`. Although, it also doesn't make sense to set
`negativeProperties` at all. This variable is only used as an auxiliary
variable to get `negativePrefix` and `negativeSuffix`. Therefore, I 
believe that's was mistake when compact support was implemented. That
mistake doesn't cause any bug, because this variable isn't used and it's 
ignored.