Docs: Fix date parts type names
* Change dayPeriod -> dayperiod in documentation

Globalize actually returns lower case `dayperiod`. But the documentation
says `dayPeriod`.

This is confusing because `Intl.DateTimeFormat`'s `formatToParts` does
return `dayPeriod` but Globalize diverges from this and returns lower
case `dayperiod`. This mistake in the documentation can lead to tricky
bugs where a developer writes code according to the documentation
checking for `dayPeriod` which doesn't work.

* Change timeZoneName -> zone in documentation

Globalize actually returns a `zone` part with zone info, not a
`timeZoneName` part.

Closes #899