Build: Add package-lock.json file
Makes CI more deterministic and quicker reproduce the same
thing locally.

Without this lock file it possible that, if an indirect dependency
has a too broad version range for one of its dependencies or if
one of its dependencies violates SemVer, that the repository will
no longer be able to pass the build, and we would not know the
last known passing state since we haven't recorded what that is.

It also makes local installs a bit quicker to get going, and
gives everyone working on it locally the same packages.