Build: Remove redundant grunt-cli
Since grunt 1.0, the CLI is automatically provided.

As the entry point is 'npm test', we don't need it to be installed
globally. Having the dependency is enough, because commands provided
by dependencies are already available to the subshell (PATH) of
commands specified in package.json.

I've also updated the dependency while at it, so to resolve
some of the npm-audit noise. Not really harmful in this context
but might help in keeping broader forward Node.js compat.