Build: Remove or mark unused variables
ESLint is able to find more unused variables and parameters than
JSHint previously. Specifically cases that are defined, then assigned,
and then never used.

I disabled the rule in the previous commit as making it pass requires
changes to actual code, not just formatting. The rule is now re-enabled
and the unused code removed or marked by underscore as required by the rule.